+++ Sources/Chroma/Tree/NewRender.swift
@@ -1,38 +1,61 @@
 struct NewRenderer: ElementWalker {
-  var currentHash: Hash = ""
+  let state: BlockState
+  var currentHash: Hash
   var orientation: Orientation
   private var groupHeight = 0
   private var groupWidth = 0
+  private var currentWidth = 0
+  private var currentHeight = 0
   /*
-    Well this is where our nested groups push and pop things off the stack.
+  Well this is where our nested groups push and pop things off the stack.
   
-    Well better we know how much space is available and consumed at this point.
-    */
+  Well better we know how much space is available and consumed at this point.
+  */
   mutating func pushNewGroup() {
-    groupHeight = height
+    groupHeight = currentHeight
     groupWidth = 0
   }
   mutating func popGroup() {
     switch orientation {
     case .horizontal:
-      width = max(groupWidth, width)
-      height = max(groupHeight + 1, height)
+      currentWidth = max(groupWidth, currentWidth)
+      currentHeight = max(groupHeight + 1, currentHeight)
     case .vertical:
-      width = max(groupWidth, width)
-      height = max(groupHeight, height)
+      currentWidth = max(groupWidth, currentWidth)
+      currentHeight = max(groupHeight, currentHeight)
     }
   }
-  private var width = 0
-  private var height = 0
   private let fg: Shell.Color = .blue
   private let bg: Shell.Color = .green
-  init(height: Int, width: Int) {
+  private var seenSelected: Bool = false
+  private let width: Int
+  private let height: Int
+
+  init(state: BlockState, width: Int, height: Int) {
+    self.height = height
+    self.width = width
+    self.state = state
     self.tiles = Array(
-      repeating: Array(repeating: Tile(symbol: "%", fg: fg, bg: bg), count: width),
+      repeating: Array(repeating: Tile(), count: width),
       count: height)
     self.orientation = .vertical
+    self.currentHash = hash(contents: "\(0)")
   }
   var tiles: [[Tile]]
+
+  var ascii: String {
+    var out = ""
+    for row in tiles {
+      var line = ""
+      for rune in row {
+        line += rune.ascii
+      }
+      line += "\n"
+      out += line
+    }
+    out.removeLast()  // remove last newline.
+    return out
+  }
 
   var _raw: String {
     var out = ""
@@ -49,6 +72,37 @@ struct NewRenderer: ElementWalker {
   }
 
   mutating func walkText(_ text: String, _ binding: InputHandler?) {
+    // This HACK needs to come before we update seenSelected for now.
+    switch orientation {
+    case .horizontal:
+      ()
+    case .vertical:
+      // HACK for vertical scrolling.
+      if groupHeight >= height {
+        if !seenSelected {
+          groupHeight -= 1  // Override the last row.
+          for (i, _) in tiles.enumerated() {
+            if i < tiles.count - 1 {
+              tiles[i] = tiles[i + 1]
+            }
+          }
+          tiles[tiles.count - 1] = Array(repeating: Tile(), count: width)
+        } else {
+          Log.error("Too many rows \(text)")
+          return
+        }
+      }
+    }
+    let fg: Shell.Color
+    let bg: Shell.Color
+    if self.state.selected == currentHash {
+      seenSelected = true
+      fg = .yellow
+      bg = .purple
+    } else {
+      fg = .blue
+      bg = .green
+    }
     switch self.orientation {
     case .horizontal:
       guard tiles[0].count >= groupWidth + text.count else {
